Added Losses to use with Extreme Weather Detection
Added Jaccard Loss and Dice Loss, two losses that are highly
relevant for image segmentation tasks like the one we are trying to
perform here. At the moment these losses are pytorch nn.module classes,
however, they may be moved around or wrapped in separate code for the
sake of compatibility with existing code.