Finalized all data preprocessing for ClimateNet
Introduced a downloader.py, splitter.py, normalizer.py,
and regridder.py module to download the ClimateNet dataset,
and preprocess it through normalization and regridding techniques.

The modules are targetted at the specific features of ClimateNet, but
they are designed to be easily modifiable for any dataset.

I also introduced a pipeline.ipynb jupyter notebook that can be used to
fine-tune ClimaX for extreme weather detection using ClimateNet
from start to finish. At the time of this commit, the pipeline.ipynb
notebook completely performs the data preprocessing.